Track referred symbols from instruction range
This is required for 'Go to references' to work for macros and show
where the macro is used. Previously `symbols.ts` only tracked references
to symbols in the data range, but when you invoke a macro it appears in
the instruction:

```assembly
  MyMacro arg1,arg2
```

One side effect of this is that 'Go to references' also works on regular
instructions and directives so for example you could view all uses of
the `movem` instruction in your project. We could prevent this but it
might be useful imo.

I did a bit of refactoring so the `ASMLine` now extracts the part of the
instruction before and after the dot, as there are now a couple of
places we need this and it seems good for completeness in parsing the
line.

So for the line ` move.w d0,d1` we now have references to:
 - instruction = `move.w` (existing)
 - mnemonic = `move`
 - size = `w`